tests-mongo/requirements.txt
fastapi
requests
python-dotenv
numpy
pytest
pytest-asyncio
motor
pymongo
httpx
pytest-mock 